+++ src/naming/beat/BeatTask.cpp
@@ -23,7 +23,7 @@ void BeatTask::run() {
         return;
     }
     uint64_t now_ms = TimeUtils::getCurrentTimeInMs();
-    _objectConfigData->_beatReactor->_delayedThreadPool->schedule(this, now_ms + _interval);
+    _objectConfigData->_beatReactor->_delayedThreadPool->schedule(this, static_cast<long>(now_ms + _interval));
     _interval = _objectConfigData->_serverProxy->sendBeat(_beatInfo);
 }
